Adopt new convention for marking CLI fns.
Some fns are intended for execution from the CLI (well, currently all of them
are, but in the future maybe some of them will be internal-only). Mark fns that
are callable from the CLI with the metadata ^{:cli true}. Then check to ensure
that a fn named on the CLI is so marked. Throw an exception if it is not.